+++ src/Octoshift/Commands/create-team/CreateTeamCommandBase.cs
@@ -53,12 +53,3 @@ public class CreateTeamCommandBase : CommandBase<CreateTeamCommandArgs, CreateTe
         AddOption(Verbose);
     }
 }
-
-public class CreateTeamCommandArgs
-{
-    public string GithubOrg { get; set; }
-    public string TeamName { get; set; }
-    public string IdpGroup { get; set; }
-    public string GithubPat { get; set; }
-    public bool Verbose { get; set; }
-}
