+++ src/Octoshift/Commands/generate-mannequin-csv/GenerateMannequinCsvCommandArgs.cs
@@ -1,0 +1,12 @@
+﻿using System.IO;
+
+namespace OctoshiftCLI.Commands;
+
+public class GenerateMannequinCsvCommandArgs
+{
+    public string GithubOrg { get; set; }
+    public FileInfo Output { get; set; }
+    public bool IncludeReclaimed { get; set; }
+    public string GithubPat { get; set; }
+    public bool Verbose { get; set; }
+}
