+++ src/Octoshift/Commands/revoke-migrator-role/RevokeMigratorRoleCommandBase.cs
@@ -57,13 +57,3 @@ public class RevokeMigratorRoleCommandBase : CommandBase<RevokeMigratorRoleComma
         AddOption(GhesApiUrl);
     }
 }
-
-public class RevokeMigratorRoleCommandArgs
-{
-    public string GithubOrg { get; set; }
-    public string Actor { get; set; }
-    public string ActorType { get; set; }
-    public string GithubPat { get; set; }
-    public bool Verbose { get; set; }
-    public string GhesApiUrl { get; set; }
-}
