+++ src/Octoshift/Commands/wait-for-migration/WaitForMigrationCommandArgs.cs
@@ -1,0 +1,9 @@
+﻿
+namespace OctoshiftCLI.Commands;
+
+public class WaitForMigrationCommandArgs
+{
+    public string MigrationId { get; set; }
+    public string GithubPat { get; set; }
+    public bool Verbose { get; set; }
+}
