+++ src/Octoshift/Commands/wait-for-migration/WaitForMigrationCommandBase.cs
@@ -54,10 +54,3 @@ public class WaitForMigrationCommandBase : CommandBase<WaitForMigrationCommandAr
         return new WaitForMigrationCommandHandler(log, githubApi);
     }
 }
-
-public class WaitForMigrationCommandArgs
-{
-    public string MigrationId { get; set; }
-    public string GithubPat { get; set; }
-    public bool Verbose { get; set; }
-}
