+++ src/ado2gh/Commands/add-team-to-repo/AddTeamToRepoCommand.cs
@@ -61,14 +61,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands
             return new AddTeamToRepoCommandHandler(log, githubApi);
         }
     }
-
-    public class AddTeamToRepoCommandArgs
-    {
-        public string GithubOrg { get; set; }
-        public string GithubRepo { get; set; }
-        public string Team { get; set; }
-        public string Role { get; set; }
-        public string GithubPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
