+++ src/ado2gh/Commands/configure-autolink/ConfigureAutoLinkCommand.cs
@@ -60,14 +60,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands
             return new ConfigureAutoLinkCommandHandler(log, githubApi);
         }
     }
-
-    public class ConfigureAutoLinkCommandArgs
-    {
-        public string GithubOrg { get; set; }
-        public string GithubRepo { get; set; }
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string GithubPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
