+++ src/ado2gh/Commands/disable-repo/DisableRepoCommand.cs
@@ -54,13 +54,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands
             return new DisableRepoCommandHandler(log, adoApi);
         }
     }
-
-    public class DisableRepoCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string AdoRepo { get; set; }
-        public string AdoPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
