+++ src/ado2gh/Commands/disable-repo/DisableRepoCommandArgs.cs
@@ -1,0 +1,10 @@
+﻿namespace OctoshiftCLI.AdoToGithub.Commands;
+
+public class DisableRepoCommandArgs
+{
+    public string AdoOrg { get; set; }
+    public string AdoTeamProject { get; set; }
+    public string AdoRepo { get; set; }
+    public string AdoPat { get; set; }
+    public bool Verbose { get; set; }
+}
