+++ src/ado2gh/Commands/inventory-report/InventoryReportCommand.cs
@@ -61,12 +61,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands
                 pipelinesCsvGeneratorService);
         }
     }
-
-    public class InventoryReportCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoPat { get; set; }
-        public bool Minimal { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
