+++ src/ado2gh/Commands/inventory-report/InventoryReportCommandArgs.cs
@@ -1,0 +1,9 @@
+﻿namespace OctoshiftCLI.AdoToGithub.Commands;
+
+public class InventoryReportCommandArgs
+{
+    public string AdoOrg { get; set; }
+    public string AdoPat { get; set; }
+    public bool Minimal { get; set; }
+    public bool Verbose { get; set; }
+}
