+++ src/ado2gh/Commands/lock-repo/LockRepoCommand.cs
@@ -54,13 +54,4 @@ namespace OctoshiftCLI.AdoToGithub.Commands
             return new LockRepoCommandHandler(log, adoApi);
         }
     }
-
-    public class LockRepoCommandArgs
-    {
-        public string AdoOrg { get; set; }
-        public string AdoTeamProject { get; set; }
-        public string AdoRepo { get; set; }
-        public string AdoPat { get; set; }
-        public bool Verbose { get; set; }
-    }
 }
