+++ src/ado2gh/Commands/share-service-connection/ShareServiceConnectionCommandArgs.cs
@@ -1,0 +1,10 @@
+﻿namespace OctoshiftCLI.AdoToGithub.Commands;
+
+public class ShareServiceConnectionCommandArgs
+{
+    public string AdoOrg { get; set; }
+    public string AdoTeamProject { get; set; }
+    public string ServiceConnectionId { get; set; }
+    public string AdoPat { get; set; }
+    public bool Verbose { get; set; }
+}
